+++ moirai-core/src/test/scala/com/nike/moirai/server/config/ProportionOfUsersConfigDeciderSpec.scala
@@ -15,6 +15,8 @@ class ProportionOfUsersConfigDeciderSpec extends FunSpec with Matchers with Gene
     val decider = new ProportionOfUsersConfigDecider[Map[String, Double]]() {
       override protected def enabledProportion(config: Map[String, Double], featureIdentifier: String): Optional[lang.Double] =
         config.get(featureIdentifier).map(Double.box).asJava
+
+      override protected def featureGroup(config: Map[String, Double], featureIdentifier: String): Optional[String] = Optional.empty()
     }
 
     val config = Map(
@@ -49,7 +51,7 @@ class ProportionOfUsersConfigDeciderSpec extends FunSpec with Matchers with Gene
       it("should return true for approximately 70% of users") {
         val results = mutable.Buffer.empty[Boolean]
 
-        implicit val generatorDrivenConfig = PropertyCheckConfiguration(minSuccessful = 10000)
+        implicit val generatorDrivenConfig: PropertyCheckConfiguration = PropertyCheckConfiguration(minSuccessful = 10000)
 
         forAll { (userId: String) =>
           results += decider.test(new ConfigDecisionInput(config, feature, FeatureCheckInput.forUser(userId)))
@@ -66,6 +68,59 @@ class ProportionOfUsersConfigDeciderSpec extends FunSpec with Matchers with Gene
         forAll { (userId: String) =>
           decider.test(new ConfigDecisionInput(config, feature, FeatureCheckInput.forUser(userId))) shouldBe false
         }
+      }
+    }
+
+    describe("featureGroup") {
+      val decider = new ProportionOfUsersConfigDecider[Map[String, Double]]() {
+        override protected def enabledProportion(config: Map[String, Double], featureIdentifier: String): Optional[lang.Double] =
+          config.get(featureIdentifier).map(Double.box).asJava
+
+        override protected def featureGroup(config: Map[String, Double], featureIdentifier: String): Optional[String] = featureIdentifier match {
+          case "feature1" => Optional.of("group1")
+          case "feature2" => Optional.of("group1")
+          case "feature3" => Optional.of("group2")
+          case "feature4" => Optional.of("group2")
+          case _ => Optional.empty()
+        }
+      }
+
+      val config = Map(
+        "feature1" -> 0.5,
+        "feature2" -> 0.5,
+        "feature3" -> 0.5,
+        "feature4" -> 0.5,
+        "feature5" -> 0.5
+      )
+
+      it("should decide features are enabled for the same users if the features are in the same feature group and have the same proportion enabled") {
+        val feature1EnabledUsers = mutable.Buffer.empty[String]
+        val feature2EnabledUsers = mutable.Buffer.empty[String]
+        val feature3EnabledUsers = mutable.Buffer.empty[String]
+        val feature4EnabledUsers = mutable.Buffer.empty[String]
+        val feature5EnabledUsers = mutable.Buffer.empty[String]
+
+        implicit val generatorDrivenConfig: PropertyCheckConfiguration = PropertyCheckConfiguration(minSuccessful = 1000)
+
+        forAll { (userId: String) =>
+          List(
+            ("feature1", feature1EnabledUsers),
+            ("feature2", feature2EnabledUsers),
+            ("feature3", feature3EnabledUsers),
+            ("feature4", feature4EnabledUsers),
+            ("feature5", feature5EnabledUsers)).foreach {
+            case (feature, buffer) if decider.test(new ConfigDecisionInput(config, feature, FeatureCheckInput.forUser(userId))) => buffer += userId
+            case _ =>
+          }
+        }
+
+        feature1EnabledUsers should contain theSameElementsAs feature2EnabledUsers
+        feature3EnabledUsers should contain theSameElementsAs feature4EnabledUsers
+
+        feature1EnabledUsers should not (contain theSameElementsAs feature3EnabledUsers)
+        feature1EnabledUsers should not (contain theSameElementsAs feature5EnabledUsers)
+
+        feature3EnabledUsers should not (contain theSameElementsAs feature5EnabledUsers)
       }
     }
   }